+++ packages/layout/examples/base-kbd.vue
@@ -1,0 +1,3 @@
+<template>
+  <span> <c-kbd>shift</c-kbd> + <c-kbd>H</c-kbd> </span>
+</template>