+++ packages/system/src/system.types.ts
@@ -1,5 +1,18 @@
-import { Component, Fragment, Suspense, Teleport } from 'vue'
-import { SystemProps, ResponsiveValue } from '@chakra-ui/styled-system'
+import {
+  Component,
+  Fragment,
+  Suspense,
+  Teleport,
+  ComponentObjectPropsOptions,
+} from 'vue'
+import {
+  SystemProps,
+  ResponsiveValue,
+  StyleProps,
+  ThemeTypings,
+} from '@chakra-ui/styled-system'
+import { IntrinsicElementAttributes } from './dom.types'
+import { Dict } from '@chakra-ui/utils'
 
 export type Tag =
   | string
@@ -8,12 +21,16 @@ export type Tag =
   | typeof Suspense
   | Component
 
-export interface ThemingProps {
-  variant?: string
-  size?: string
-  colorScheme?: string
+export interface ThemingProps<ThemeComponent extends string = string> {
+  variant?: ThemeComponent extends keyof ThemeTypings['components']
+    ? ThemeTypings['components'][ThemeComponent]['variants'] | (string & {})
+    : string
+  size?: ThemeComponent extends keyof ThemeTypings['components']
+    ? ThemeTypings['components'][ThemeComponent]['sizes'] | (string & {})
+    : string
+  colorScheme?: ThemeTypings['colorSchemes'] | (string & {})
   orientation?: 'vertical' | 'horizontal'
-  styleConfig?: any
+  styleConfig?: Dict
 }
 
 export interface ChakraProps extends SystemProps {
@@ -46,4 +63,27 @@ export interface ChakraProps extends SystemProps {
   noOfLines?: ResponsiveValue<number>
 }
 
-export type { SystemStyleObject } from '@chakra-ui/styled-system'
+type ElementType<P = any> =
+  | {
+      [K in keyof IntrinsicElementAttributes]: P extends IntrinsicElementAttributes[K]
+        ? K
+        : never
+    }[keyof IntrinsicElementAttributes]
+  | Component<P>
+
+export type As<Props = any> = ElementType<Props>
+
+/**
+ * Extract the props of a Vue element or component
+ */
+export type PropsOf<T extends As> = T & {
+  as?: As
+}
+
+export type HTMLChakraProps<T extends As> = Omit<
+  PropsOf<T>,
+  T extends 'svg'
+    ? 'ref' | 'children' | keyof StyleProps
+    : 'ref' | keyof StyleProps
+> &
+  ChakraProps & { as?: As }