+++ tooling/auto-import/src/index.ts
@@ -1,9 +1,10 @@
-import kebabCase from 'lodash.kebabcase'
+import * as ChakraComponenets from '@chakra-ui/vue-next'
 
 export const componentResolver = (name: string) => {
-  if (kebabCase(name).startsWith('c-'))
+  if (name in ChakraComponenets) {
     return {
       importName: name,
       path: `@chakra-ui/vue-next`,
     }
+  }
 }