+++ website/src/components/home/Features.vue
@@ -1,0 +1,8 @@
+<template>
+  <chakra.div>
+    <c-heading>What's coming to v1?</c-heading>
+  </chakra.div>
+  <chakra.section> </chakra.section>
+</template>
+
+<script setup lang="ts"></script>