+++ website/src/components/home/PilersHero.vue
@@ -3,19 +3,17 @@
     as="section"
     :flex-dir="['column', 'column', 'row']"
     font-family="body"
-    :mx="[0, 10, 12]"
     :py="[4, 10, 10]"
   >
-    <chakra.div
-      d="grid"
-      :grid-template-columns="['auto', 'auto', 'repeat(3, 1fr)']"
-      :style="{ gap: '1.5rem' }"
+    <c-grid
+      :template-columns="['auto', 'auto', 'repeat(3, 1fr)']"
+      gap="1.5rem"
       shadow="lg"
       bg="white"
       py="10"
       :px="{ base: '3', sm: '10' }"
     >
-      <CBox m="3">
+      <CBox my="3" :mx="{ base: '0', sm: '3' }">
         <CFlex align-items="center" mb="5">
           <CFlex
             justify-content="center"
@@ -28,12 +26,12 @@
           >
             <IconAccessibilityAlt />
           </CFlex>
-          <chakra.h2 font-size="md"> Accessible </chakra.h2>
+          <chakra.h2 font-size="xl"> Accessible </chakra.h2>
         </CFlex>
         Chakra UI strictly follows WAI-ARIA standards. All components come with
         proper attributes and keyboard interactions out of the box.
       </CBox>
-      <CBox m="3">
+      <CBox my="3" :mx="{ base: '0', sm: '3' }">
         <CFlex align-items="center" mb="5">
           <CFlex
             justify-content="center"
@@ -46,12 +44,12 @@
           >
             <ColorPalette />
           </CFlex>
-          <chakra.h2 font-size="md"> Themeable </chakra.h2>
+          <chakra.h2 font-size="xl"> Themeable </chakra.h2>
         </CFlex>
         Quickly and easily reference values from your theme throughout your
         entire application, on any component.
       </CBox>
-      <CBox m="3">
+      <CBox m="3" :mx="{ base: '0', sm: '3' }">
         <CFlex align-items="center" mb="5">
           <CFlex
             justify-content="center"
@@ -64,12 +62,12 @@
           >
             <Cube />
           </CFlex>
-          <chakra.h2 font-size="md"> Composable </chakra.h2>
+          <chakra.h2 font-size="xl"> Composable </chakra.h2>
         </CFlex>
         Components were built with composition in mind. You can leverage any
         component to create new things.
       </CBox>
-    </chakra.div>
+    </c-grid>
   </CFlex>
 </template>