+++ packages/core/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @chakra-ui/vue-next
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- [`0b34bda`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/0b34bda9396859324e7c90a1bfea6483f1022585) Thanks [@codebender828](https://github.com/codebender828)! - Fixes dependency resolution for code and layout components"
+
+- Updated dependencies [[`ad7d2f2`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/ad7d2f2db3bcce5d278fe73384d5c8bfae5503a9)]:
+  - @chakra-ui/c-code@1.0.0-alpha.2
+  - @chakra-ui/vue-layout@0.1.0-alpha.3
+
+## 1.0.0-alpha.3
+
+### Minor Changes
+
+- [#61](https://github.com/chakra-ui/chakra-ui-vue-next/pull/61) [`520550f`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/520550f0a233af45d83a5dd49907646422b39eb9) Thanks [@koca](https://github.com/koca)! - Created layout components and updated theminnng types for packages.
+
+### Patch Changes
+
+- [#59](https://github.com/chakra-ui/chakra-ui-vue-next/pull/59) [`a79204f`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/a79204f8a6786fdb6456632350480a14e17ad345) Thanks [@devivekw](https://github.com/devivekw)! - Created code component
+
+- Updated dependencies [[`a79204f`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/a79204f8a6786fdb6456632350480a14e17ad345), [`520550f`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/520550f0a233af45d83a5dd49907646422b39eb9), [`e65c3b4`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/e65c3b402a3e795ff32b81e9bf00be5bcbc739cb)]:
+  - @chakra-ui/c-accordion@0.1.0-alpha.2
+  - @chakra-ui/c-alert@1.0.0-alpha.2
+  - @chakra-ui/c-button@1.0.0-alpha.2
+  - @chakra-ui/c-close-button@0.1.0-alpha.2
+  - @chakra-ui/c-color-mode@0.1.0-alpha.2
+  - @chakra-ui/c-flex@1.0.0-alpha.2
+  - @chakra-ui/c-icon@1.0.0-alpha.2
+  - @chakra-ui/c-modal@1.1.0-alpha.1
+  - @chakra-ui/c-popper@0.1.0-alpha.2
+  - @chakra-ui/c-portal@0.1.0-alpha.2
+  - @chakra-ui/c-reset@1.0.0-alpha.2
+  - @chakra-ui/c-spinner@1.0.0-alpha.2
+  - @chakra-ui/c-theme-provider@1.0.0-alpha.2
+  - @chakra-ui/c-visually-hidden@1.0.0-alpha.2
+  - @chakra-ui/vue-system@0.1.0-alpha.2
+  - @chakra-ui/vue-theme@0.1.0-alpha.2
+  - @chakra-ui/vue-theme-tools@0.1.0-alpha.2
+
 ## 1.0.0-alpha.2
 
 ### Minor Changes